+++ CHANGELOG.md
@@ -2,9 +2,19 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [41.1.0] 2026-03-31
+## [41.2.0] 2026-04-02
 ### Added
 - Support for targetConstraint extension.
+
+## [41.1.1] 2026-04-02
+### Fixed
+- Original score extensions (ordinalValue or itemWeight) were not
+  kept in the exported questionnaire.
+
+## [41.1.0] 2026-03-30
+### Added
+- Support for rendering-markdown extension. Note that rendering-xhtml
+  takes precedence over rendering-markdown if both are present.
 
 ## [41.0.0] 2026-03-27
 ### Changed